Add weather data retrieval functionality
This commit adds the functionality to retrieve weather data from the OpenWeatherMap API. It includes the ability to fetch current weather information and a 5-day forecast for one or two cities based on user input. The data is parsed and displayed in the Django template. This functionality is part of the Weather app in our project.